pyproject.toml
[project]
name = "easygo-tracker"
version = "1.0.0"
description = "Telegram Bot"
requires-python = ">=3.11,<4.0"
dependencies = [
    # Telegram Bot
    "python-telegram-bot[ext]==20.7",

    # MongoDB and ODM
    "motor==3.4.0",
    "beanie==1.25.0",
    "pymongo==4.6.3",

    # Configuration Management
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",

    # Logging
    "structlog==24.1.0",

    # Retry Logic
    "tenacity==8.2.3",

    # Utilities
    "python-dateutil==2.8.2",

    # HTTP Client
    "aiohttp==3.9.1",
    "requests==2.31.0",

    # Google Sheets
    "gspread>=6.1.0",
    "google-auth>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]
